[BUG-FIX]: extra new line
In some cases the `buffer` may exceed the allowed `width`. When this
happens, the closing tag may still be in the buffer and it just may
happen that the place where the `buffer` will be cut is at the end of
the closing tag. This makes the writer print the new line character.
Now we are at a position in which we need to decide if we want to print
another new line. We wish to do this only if the node is a `BLOCK` node
(but with the extra condition added to the code). The `li` element has
also been added to the `BLOCK` list.