[NEW]: add and delete from BLOCK
This writing style has several elements that are treated as blocks.
That is, we display them in a separate line. Sometimes however, you may
wish to have some other elements treated as blocks, or perhaps you
don't wish certain element in the defaults to be treated as a block.
For this reason I have added the options `add_block` and `del_block`.
When using the command line use the form `@add_block=[ele1,ele2]` to
have your elements added to the list, or `@del_block=[ele1,ele2]` to
remove them from it. Note that regardless of the order in which you
specify the options, the `pre_process` function in this style will
first perform the appending of the elements and then the deletion of
the elements.