[SYNTAX]: blocks go in a new line
By removing the split keyword, we can avoid having the writer split the
line at the beginning of a new element. Consider
`<em>i</em><strong>-th</strong>`. If it would happen that the line
would not fit in the allowed width then the writer would see that we
can split the line at `<strong`. This however is wrong because a new
line character is considered a space and thus we would have `<em>i</em>
<strong>-th</strong>`. Thus the decision to remove the break hint. We
may need to go back to this line and added back again if we can locate
some elements in which it would be safe to split.